refactoring login view
