+++ src/views/Detail.vue
@@ -32,7 +32,7 @@
 
 <script>
 import axios from 'axios'
-import Header from "../components/Header"
+import Header from "../components/Organisms/Header"
 export default {
   props: ["id"],
   data() {
