+++ src/views/Profile.vue
@@ -39,7 +39,7 @@
 
 <script>
 import axios from 'axios'
-import Header from "../components/Header"
+import Header from "../components/Organisms/Header"
 export default {
   data() {
     return {
